orchestrator/requirements.txt
requests>=2.31.0
streamlit>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
pandas>=2.0.0
pytest>=7.0.0
